fix: try different sprintf format string for R_xlen_t

+++ src/checks.c
@@ -486,15 +486,15 @@ SEXP attribute_hidden c_check_integerish(SEXP x, SEXP tol, SEXP lower, SEXP uppe
                 }
                 break;
             case INT_RANGE:
-                snprintf(msg, 255, "Must be of type 'integerish', but element %jd is not in integer range", ok.pos);
+                snprintf(msg, 255, "Must be of type 'integerish', but element %g is not in integer range", (double)ok.pos);
                 return ScalarString(mkChar(msg));
                 break;
             case INT_TOL:
-                snprintf(msg, 255, "Must be of type 'integerish', but element %jd is not close to an integer", ok.pos);
+                snprintf(msg, 255, "Must be of type 'integerish', but element %g is not close to an integer", (double)ok.pos);
                 return ScalarString(mkChar(msg));
                 break;
             case INT_COMPLEX:
-                snprintf(msg, 255, "Must be of type 'integerish', but element %jd has an imaginary part", ok.pos);
+                snprintf(msg, 255, "Must be of type 'integerish', but element %g has an imaginary part", (double)ok.pos);
                 return ScalarString(mkChar(msg));
                 break;
         }
